more refactoring and tests

+++ src/log/source_location.h
@@ -42,8 +42,8 @@ namespace tasarch::log {
           source_location __loc;
           __loc._M_file = __file;
           __loc._M_func = __func;
-          __loc._M_line = __line;
-          __loc._M_col = __col;
+          __loc._M_line = static_cast<uint_least32_t>(__line);
+          __loc._M_col = static_cast<uint_least32_t>(__col);
           return __loc;
         }
